Fix Bunch O' Tallow Candle Placing
Tallow candles now correctly place a "bunchotallowcandles" block instead of a normal "bunchocandles" block.

This allows for unique properties between the two types of candles and also prevents an exploit where a player could place tallow candles as a "bunchocandles" and then break that bunch to receive normal wax candles.